demo-python/requirements.txt
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-deepseek>=0.1.0
langchain-tavily>=0.1.0
colorama>=0.4.6
psutil>=5.9.0 